fix go1.18
