+++ go.mod
@@ -1,3 +1,3 @@
 module github.com/modern-go/reflect2
 
-go 1.18
+go 1.12
